do not overwrite existing files in init_nlo script

+++ COMIX/Main/init_nlo.sh
@@ -42,6 +42,7 @@ for i in $nt/Process/Comix/*.map; do
     mv $i.tmp $(echo $i | sed -e's/\(__NQ_.*\)[.]map/.map\1/g' \
       -e's/.map/__QCD('$tp').map/g' -e's/.map\(.*\)/\1.map/g');
   else
+    test -f $(echo $i | sed -e's/.map/__QCD('$tp').map/g') && continue;
     sed -e'1 s/ /__QCD('$tp') /g' -e'1 s/$/__QCD('$tp')/g' $i > $i.tmp;
     if awk '{ if ($1!=$2) exit 1; exit 0; }' < $i; then rm $i.tmp;
     else mv $i.tmp $(echo $i | sed -e's/.map/__QCD('$tp').map/g'); fi;
